separated database.cs into tags and database
